+++ launchers/data-loader-cli/src/test/java/org/eclipse/dataspaceconnector/examples/JsonFileGenerator.java
@@ -55,8 +55,8 @@ public class JsonFileGenerator {
         return ContractDefinition.Builder.newInstance()
                 .id("ContractDefinition_" + i)
                 .selectorExpression(AssetSelectorExpression.SELECT_ALL)
-                .contractPolicy(Policy.Builder.newInstance().id("CP_" + i).build())
-                .accessPolicy(Policy.Builder.newInstance().id("AP_" + i).build())
+                .contractPolicyId("CP_" + i)
+                .accessPolicyId("AP_" + i)
                 .build();
     }